Fix performance issue on artifacts with workflow
Management of workflow since the introduction of artifact based
post actions is slow.
A basic workflow on 4 values (4x4 transitions) can lead to 1900+
SQL queries at display and 800+ queries at write (artifact update).

This commit solves this perf issue with 2 strategies:
- caching of the already computed values. Some may argue it hides the
  underlying problem but given the current architecture of the tracker
  it's hard to do it without an entire rebuild of the way fields are
  managed.
- pre-fetching of the values. The code is desgined after fetching of
  single transitions. However, most of the time we need to fetch the
  entire worfklow (all transitions, all post actions, etc) so it's
  more efficient to fetch everything at once.

With this patch, an artifact web view display went from 1870 SQL queries
to 299 (for the record, same artifact without workflow requires 278 SQL
queries) and update from 311 to 150.

Fix request #17191 Artifacts in trackers with workflow generate way to much SQL requests

+++ plugins/tracker/include/Workflow/WorkflowFactory.class.php
@@ -51,6 +51,10 @@ class WorkflowFactory // phpcs:ignore PSR1.Classes.ClassDeclaration.MissingNames
      * @var StateFactory
      */
     private $state_factory;
+    /**
+     * @var array
+     */
+    private $cache;
 
     /**
      * Should use the singleton instance()
@@ -81,22 +85,17 @@ class WorkflowFactory // phpcs:ignore PSR1.Classes.ClassDeclaration.MissingNames
      */
     protected static $_instance; // phpcs:ignore PSR2.Classes.PropertyDeclaration.Underscore
 
-    public static function setInstance($instance)
+    public static function setInstance(self $instance): void
     {
         self::$_instance = $instance;
     }
 
-    public static function clearInstance()
+    public static function clearInstance(): void
     {
         self::$_instance = null;
     }
 
-    /**
-     * The singleton method
-     *
-     * @return WorkflowFactory
-     */
-    public static function instance()
+    public static function instance(): self
     {
         if (! isset(self::$_instance)) {
             $formelement_factory = Tracker_FormElementFactory::instance();
@@ -131,22 +130,12 @@ class WorkflowFactory // phpcs:ignore PSR1.Classes.ClassDeclaration.MissingNames
         return self::$_instance;
     }
 
-    /**
-     * @return Tracker_Workflow_Trigger_RulesManager
-     */
-    public function getTriggerRulesManager()
+    public function getTriggerRulesManager(): Tracker_Workflow_Trigger_RulesManager
     {
         return $this->trigger_rules_manager;
     }
 
-    /**
-     * Build a Workflow instance
-     *
-     * @param array $row The data describing the workflow
-     *
-     * @return Workflow
-     */
-    public function getInstanceFromRow($row)
+    public function getInstanceFromRow(array $row): Workflow
     {
         $tracker = $this->tracker_factory->getTrackerById($row['tracker_id']);
         if ($tracker === null) {
@@ -166,24 +155,22 @@ class WorkflowFactory // phpcs:ignore PSR1.Classes.ClassDeclaration.MissingNames
         );
     }
 
-    /**
-     * @param int $workflow_id
-     * @return null|Workflow
-     */
-    public function getWorkflow($workflow_id)
+    public function getWorkflow(int $workflow_id): ?Workflow
     {
-        $row = $this->getDao()->searchById($workflow_id);
-        if ($row !== null) {
-            return $this->getInstanceFromRow($row);
+        if (! isset($this->cache[$workflow_id])) {
+            $this->cache[$workflow_id] = false;
+            $row = $this->getDao()->searchById($workflow_id);
+            if ($row !== null) {
+                $this->cache[$workflow_id] = $this->getInstanceFromRow($row);
+            }
         }
-        return null;
+        if ($this->cache[$workflow_id] === false) {
+            return null;
+        }
+        return $this->cache[$workflow_id];
     }
 
-
-    /**
-     * @return WorkflowWithoutTransition
-     */
-    public function getWorkflowWithoutTransition(Tracker $tracker)
+    public function getWorkflowWithoutTransition(Tracker $tracker): WorkflowWithoutTransition
     {
         return new WorkflowWithoutTransition(
             $this->getGlobalRulesManager($tracker),
@@ -237,7 +224,7 @@ class WorkflowFactory // phpcs:ignore PSR1.Classes.ClassDeclaration.MissingNames
     public function deleteWorkflow($workflow_id)
     {
         $workflow = $this->getWorkflow($workflow_id);
-        if ($this->transition_factory->deleteWorkflow($workflow)) {
+        if ($workflow && $this->transition_factory->deleteWorkflow($workflow)) {
             return $this->delete($workflow_id);
         }
     }
